Change chimney size on wish list page - issue #5.

+++ src/Web/Controllers/AccountController.cs
@@ -32,7 +32,7 @@ namespace Wishes.Web.Controllers
                                                Address = model.Address,
                                                Zipcode = model.Zipcode,
                                                City = model.City,
-                                               ChimneySize = (int)ChimneySize.Medium
+                                               ChimneySize = ChimneySize.Medium
                                            });
 
                 FormsAuthentication.SetAuthCookie(model.Username, true);
